debug: fix collisions (a little) by fixing quadtree ::update
There was a problem in the DFT system in CollisionSystem::update.
Essentially, it was exiting too early, and discarding the lowest layer
of nodes - something like that - and after some pen and paper debugging,
I came up with a slight change that fixed it. As a result, collisions
now work as expected, but are still jank. Needs more work.