physics, #1: fixed unfound ents
the force search method was incorrectly returning nullpointer even if
the entity was found; I discovered this by adding more logs and inspecting
the logs - the "unfound" entity was still in the tree. It was an easy fix
after the correct diagnosis.

There's another issue about WHY these ents aren't in the right node to
begin with.