ecs: instead of using strings in the component map, use ints
I realized that ints were probably faster to hash / compare etc; so
I replaced the typename keys with a basic templating int key.

The benefit of the templating is that it'll complain if you try to
use (i.e. call get<> etc.) with a component that hasn't been declared
a value for; because the compiler can't compile it. So you're guaranteed
to at least have a value for each component.